refactor deals datasource stuff (#102)
