+++ src/dealWizard/makeOfferWizard/makeOfferWizardManager.ts
@@ -1,9 +1,8 @@
 import { autoinject, useView } from "aurelia-framework";
 import { PLATFORM } from "aurelia-pal";
 import { Router, RouterConfiguration } from "aurelia-router";
-import { IDealRegistrationData } from "entities/Deal";
+import { DealRegistrationData, IDealRegistrationData } from "entities/DealRegistrationData";
 import { WizardService, IWizardState, IWizardStage } from "../../services/WizardService";
-import { RegistrationData } from "../registrationData";
 
 @useView(PLATFORM.moduleName("../wizardManager.html"))
 @autoinject
@@ -41,7 +40,7 @@ export class MakeOfferWizardManager {
     route: "stage4",
     moduleId: PLATFORM.moduleName("../stages/partnerDaoStage/partnerDaoStage"),
   }];
-  private registrationData = new RegistrationData();
+  private registrationData = new DealRegistrationData();
 
   constructor(public wizardService: WizardService) {
     this.wizardState = this.wizardService.registerWizard(this, this.stages, this.registrationData);
