+++ src/main.ts
@@ -16,7 +16,7 @@ import { TimingService } from "services/TimingService";
 import { TokenService } from "services/TokenService";
 import { CeramicServiceMock } from "services/CeramicServiceMock";
 import { Deal } from "entities/Deal";
-import { DataSourceDeals } from "services/DataSourceDeals";
+import { IDataSourceDeals } from "services/DataSourceDealsTypes";
 
 export function configure(aurelia: Aurelia): void {
   aurelia.use
@@ -29,7 +29,7 @@ export function configure(aurelia: Aurelia): void {
       configuration.settings.keyboard = false;
     });
   aurelia.use.singleton(HTMLSanitizer, DOMPurify);
-  aurelia.use.singleton(DataSourceDeals, CeramicServiceMock);
+  aurelia.use.singleton(IDataSourceDeals, CeramicServiceMock);
 
   const network = process.env.NETWORK as AllowedNetworks;
   const inDev = process.env.NODE_ENV === "development";
